# Add `routes --unused` option to detect extraneous routes.