Routes take a long time to draw. Over time, a Rails app can become slow to boot simply because of how many routes it has. This script can be used to detect routes that are drawn, but aren't actually valid. Removing routes this script detects can help speed up your app and remove dead code.

Example:

```
> bin/rails routes --unused

Found 2 unused routes:

Prefix Verb URI Pattern    Controller#Action
   one GET  /one(.:format) action#one
   two GET  /two(.:format) action#two
```

I'm having trouble deciding if this is worthy of upstreaming to Rails, or if it should be a gem. It seems like something many large apps can benefit from, and uses private API, so I believe placing it here is the best fit. What do you think?

For extra context, I wrote this script for a very large app after noticing it had some unused routes. It found 100+ invalid routes I'm currently working through removing. It has unveiled a lot of dead code and I think it would be useful to share with other apps.