@dhh 2022-07-30
Could you share some performance figures from what cutting 100 routes on this very large app was able to do to boot time? Curious to get a sense of scale here.