@lsylvester 2022-07-30
Another place that this could live as a rubocop rule - which would make it easier to prevent the unused routes from accumulating without adding additional steps to CI to check for unused routes. 

If this requires using private APIs - it might be worth exposing the information it requires via public apis so that the a rule could be written outside of Rails.