@ghiculescu 2022-08-01
> Another place that this could live as a rubocop rule

I'm not sure if that could work, since this requires Rails to boot, but rubocop runs statically.