@xjunior 2022-08-01
Very useful. I just felt the name is misleading (`unused` => `invalid`).