@simi 2022-08-01
:thinking: Would it be possible to use routes namespace and make command `bin/rails routes:invalid` or so?