@simi 2022-08-01
> The main reason I'm skeptical here is that "unused" routes is in part a design feature.

Alternatively it could be called "not implemented" routes. That could be helpful for development to see what's missing.