@gmcgibbon 2022-08-02
After removing 66 of the 120 unused routes from the app, I ran a boot profile. The difference isn't huge, but I think it illustrates an interesting picture.

Before (`1.45s` to draw with 120 unused routes):

<img width="582" alt="Screen Shot 2022-08-01 at 11 43 01 PM" src="https://user-images.githubusercontent.com/5162312/182295836-515a873b-31de-4b5a-b792-0e06dc067eaf.png">

After (`1.39s` to draw with 66 less unused routes):

<img width="446" alt="Screen Shot 2022-08-01 at 11 43 14 PM" src="https://user-images.githubusercontent.com/5162312/182295888-46dafd29-e826-4422-82ee-69e481a702ee.png">

If we assume a roughly 2x improvement from the initial results I've profiled, that's 60ms * 2 is 120ms. So that's ~1ms per route? I was mainly doing this to stop consumers of an API from hitting a potentially unwanted route, but there is _some_ performance merit to doing this as well depending on how many routes the script finds. It is less impactful than I assumed, but still worth looking at to cut down on boot time in an app IMO.

> You could start with just index, then later add show, then create, etc. All the while never needing to change the routes.rb file. That's a pretty key DX feature to me.

This is a fair point, and I believe I've done this in the past. It may depend on the developer's intentions, but 95% of the use-cases I've seen are people removing actions/views and forgetting to update the routes file. The rest have by majority been resource route definitions with extraneous routes that never got implemented.

Note that I used speedscope and stackprof in wall mode to generate the profiles.