@gmcgibbon 2022-08-02
> Very useful. I just felt the name is misleading (unused => invalid).

> Alternatively it could be called "not implemented" routes.


I think I'm a little inconsistent on the wording in the script. I'll update it to be `unused` everywhere as invalid routes would raise an exception. Thanks for the feedback!