@gmcgibbon 2022-08-02
The script has been updated to be invoked via `bin/rails rails --unused`, supports filtering, and also exits with a nonzero status code when it finds extra routes.