@gmcgibbon 2022-08-05
I don't see much opposition to the general idea of the script, so I'm going to merge this. Happy to revert if we find that it isn't valuable enough to be in rails/rails and I can make a gem.