@natematykiewicz 2022-08-05
> The main reason I'm skeptical here is that "unused" routes is in part a design feature. It means that you can just do `resources :posts`, and then you don't have to touch your routes.rb file as you add additional methods to the PostsController. You could start with just index, then later add show, then create, etc. All the while never needing to change the routes.rb file. That's a pretty key DX feature to me.
> 
> Now it's possible that on very large codebases this feature isn't worth the performance trade-off. But we should measure that as we would any other performance optimization to ensure that expanding the command footprint is paid off.

For what it's worth, I had written a script that worked similarly a few months back (never made a PR of it). I found 4 unused routes in ActionMailbox. https://github.com/rails/rails/pull/45230